Completed work for day 7 part 1.
Too many commits in one. Micro-commiting strategy off-track.
Tests added for counting up the bags that can hold shiny gold ones.
Overall puzzle failed due to lack of handling of rules with more than
three bags. Tests for four bag rules added and functions changed and
or fixed to cater for this. Finally the puzzle was updated and run successfully.