Test for counting bags added, implemented and part 2 puzzle solved.
Very long rabbit hole trying to get the recursive count-all-bags
function to work properly. Used the wrong for/ function and then did
not set the terminating condition for an empty list correctly. Not
sure how I could have worked this through using tests. Although the
function was worked out in the REPL so possibly it could be be done
with lots of small tests. REPL is easier though. Can't believe I got
so stuck, but great practice for recursive thinking.